requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
pandas>=2.0.0
numpy>=1.24.0
streamlit>=1.28.0
google-generativeai>=0.3.0

